Refactor: 필요없는 주석 제거#50
