+++ src/hooks/useCabinetData.tsx
@@ -20,7 +20,6 @@ interface cabinetApiData {
 export const useCabinetData = (
   selectedBuilding: { name: string } | null,
   selectedFloor: number | null,
-  // selectedCabinet: { cabinetId: number; cabinetNumber: number } | null,
   isMyCabinet: boolean
 ) => {
   const [cabinetData, setCabinetData] = useState<cabinetApiData[]>([]);
@@ -38,12 +37,6 @@ export const useCabinetData = (
   };
 
   useEffect(() => {
-    // cabinetData 상태 업데이트 (사물함 버튼 관련)
-    // setCabinetData((prevData) =>
-    //   prevData.map((cabinet) =>
-    //     cabinet.id === selectedCabinet?.cabinetId ? { ...cabinet, isMine: isMyCabinet } : cabinet
-    //   )
-    // );
     if (selectedBuilding !== null && selectedFloor !== null) {
       fetchCabinetData(selectedBuilding.name, selectedFloor);
     }
