+++ src/hooks/useSearchResultButton.tsx
@@ -4,12 +4,10 @@ import { useEffect, useState } from "react";
 import { useNavigate } from "react-router";
 import { cabinetCallApi } from "@/api/cabinetCallApi";
 import { cabinetDetailInfoApi } from "@/api/cabinetDetailInfoApi";
-import { useSearch } from "@/hooks/useSearch";
 
 export const useSearchResultButton = () => {
   const navigate = useNavigate();
   const [filteredCabinetDetail, setFilteredCabinetDetail] = useState(null); // 상세 데이터 저장
-  const { setSearchParams } = useSearch();
 
   // 검색 결과 버튼 클릭 시 사물함 API 호출
   const fetchClickResultButton = async (
@@ -36,24 +34,7 @@ export const useSearchResultButton = () => {
       } else {
       }
     } catch (error) {
-      console.error("Error fetching cabinet details:", error);
-    }
-  };
-  // 수정 필요
-  const fetchSearchResultCabinetData = async (
-    building: string,
-    floor: number
-  ) => {
-    try {
-      const response = await cabinetCallApi(building, floor);
-      setSearchParams({ building, floor: floor.toString() }); // 쿼리스트링
-      console.log(200);
-
-      return response.data;
-    } catch (error) {
-      if (error === 404) {
-        console.error(404);
-      }
+      console.error(error);
     }
   };
 
@@ -72,6 +53,5 @@ export const useSearchResultButton = () => {
   return {
     fetchClickResultButton,
     filteredCabinetDetail,
-    fetchSearchResultCabinetData,
   };
 };
